+++ fiftyone/core/fields.py
@@ -15,6 +15,7 @@ import eta.core.utils as etau
 
 import fiftyone.core.utils as fou
 import fiftyone.core.frame_utils as fofu
+from fiftyone.core.odm import DictField, Field
 
 
 def parse_field_str(field_str):
@@ -46,13 +47,6 @@ def parse_field_str(field_str):
             raise ValueError("Failed to parse field string '%s'" % field_str)
 
     return ftype, embedded_doc_type, subfield
-
-
-class Field(mongoengine.fields.BaseField):
-    """Base class for :class:`fiftyone.core.sample.Sample` fields."""
-
-    def __str__(self):
-        return etau.get_class_name(self)
 
 
 class ObjectIdField(mongoengine.ObjectIdField, Field):
@@ -197,36 +191,6 @@ class PolylinePointsField(ListField):
             )
 
 
-class DictField(mongoengine.DictField, Field):
-    """A dictionary field that wraps a standard Python dictionary.
-
-    If this field is not set, its default value is ``{}``.
-
-    Args:
-        field (None): an optional :class:`Field` instance describing the type
-            of the values in the dict
-    """
-
-    def __init__(self, field=None, **kwargs):
-        if field is not None:
-            if not isinstance(field, Field):
-                raise ValueError(
-                    "Invalid field type '%s'; must be a subclass of %s"
-                    % (type(field), Field)
-                )
-
-        super().__init__(field=field, **kwargs)
-
-    def __str__(self):
-        if self.field is not None:
-            return "%s(%s)" % (
-                etau.get_class_name(self),
-                etau.get_class_name(self.field),
-            )
-
-        return etau.get_class_name(self)
-
-
 class VectorField(mongoengine.fields.BinaryField, Field):
     """A one-dimensional array field.
 
@@ -284,28 +248,6 @@ class ArrayField(mongoengine.fields.BinaryField, Field):
     def validate(self, value):
         if not isinstance(value, (np.ndarray, Binary)):
             self.error("Only numpy arrays may be used in an array field")
-
-
-class IntDictField(DictField):
-    def to_mongo(self, value):
-        if value is None:
-            return None
-
-        value = {str(k): v for k, v in value.items()}
-        return super().to_mongo(value)
-
-    def to_python(self, value):
-        if value is None:
-            return None
-
-        return {int(k): v for k, v in value.items()}
-
-    def validate(self, value):
-        if not len(value):
-            return
-
-        if any(map(lambda k: isinstance(k, six.integer_types), value)):
-            self.error("Not all keys are integers")
 
 
 class EmbeddedDocumentField(mongoengine.EmbeddedDocumentField, Field):
